Move audio player logic to a hook
