+++ app/components/Player/index.tsx
@@ -1,12 +1,8 @@
-import React, {
-  useEffect,
-  useRef,
-  useState,
-  useContext,
-  MutableRefObject,
-} from 'react';
+import React, { useContext } from 'react';
 import { Link } from '@remix-run/react';
+import { useAudio } from '~/hooks/useAudio';
 import { PlayerContext } from '~/context/playerContext/playerContextProvider';
+import { ProfileContext } from '~/context/profileContext/profileContextProvider';
 import { IconButton } from '~/components/common/Button';
 import EntityThumbnail from '~/components/Entity/EntityThumbnail';
 import { API_URLS } from '~/constants/api';
@@ -14,44 +10,24 @@ import ProgressBar from './ProgressBar';
 
 const Player = () => {
   const {
+    audioRef,
+    playPause,
+    onTimeUpdate,
+    progress,
+    setProgress,
+  } = useAudio();
+  const {
     current,
     isPlaying,
-    setIsPlaying,
   } = useContext(PlayerContext);
-  const audioRef = useRef() as MutableRefObject<HTMLAudioElement>;
-  const [progress, setProgress] = useState(0);
-
-  const playPause = () => {
-    if (isPlaying) {
-      audioRef.current.pause();
-    } else {
-      audioRef.current.play();
-    }
-    setIsPlaying(!isPlaying);
-  };
-
-  const onTimeUpdate = () => {
-    setProgress(audioRef.current.currentTime);
-  };
-
-  useEffect(() => {
-    if (progress !== audioRef.current.currentTime) {
-      if (Math.abs(Math.round(progress) - Math.round(audioRef.current.currentTime)) > 2) {
-        audioRef.current.currentTime = progress;
-      }
-    }
-  }, [progress]);
+  const { info } = useContext(ProfileContext);
 
   return (
-    <section className={`component player ${current ? 'playing' : ''}`}>
+    <section className={`component player ${current && info.address ? 'playing' : ''}`}>
       <section className="playingMusic">
-        {
-          current?.id && (
-            <Link to={`/album/${current?.id}`}>
-              <EntityThumbnail data={current} className="circle" />
-            </Link>
-          )
-        }
+        <Link to={`/album/${current?.id ?? ''}`}>
+          { current && <EntityThumbnail data={current} /> }
+        </Link>
         <header>
           <h5>{ current?.name ?? '...' }</h5>
           <span>{ current?.artistName ?? '...' }</span>
